Added "data.fonts" support for Emerald.
The table in EM is formatted differently than in RS. I'm going to make an "allFonts" [[List]] soon for Ruby, Sapphire, and Emerald.

There is 1 tab character before the first offset in some of the rows of tableReference.txt (like on lines 784 to 788), which may require revamping the file altogether.