New ViewModel for AngleComboBox
This new viewmodel includes functionality for filtering and keyboard selection (down, up, confirm) built in. It encapsulates only the combobox logic, not the logic for how the elements are found in the PokemonModel, so it's much simpler than the previous implementation.

Currently the new model is only being used for the table tool, but we can expand to use it in other places.