A bunch of feature requests. (I thought I uploaded this already.) (#150)
* Updated the "graphics.text.importer" formatting.

The "a" field was actually supposed to be 4 one-byte fields.

* Added the table for Japanese bold glyphs in EM.

* Added support for the keypad characters image.

I don't know what palette is associated with this image, though.

* Moved new table to a more appropriate location.

* Added "data.fonts" support for Emerald.

The table in EM is formatted differently than in RS. I'm going to make an "allFonts" [[List]] soon for Ruby, Sapphire, and Emerald.

There is 1 tab character before the first offset in some of the rows of tableReference.txt (like on lines 784 to 788), which may require revamping the file altogether.

* Started the [[List]] of all fonts for Emerald.

A few notes:
− "short2" comes before "short" because of the naming conventions HMA already has in place for the two short fonts that are similar but have different lower-case Ls. (short2 has the lower-case L that's used in battles, etc.). The regular "short" is "short_narrow" in the decomps.
− I suspect that the "?" entries are Japanese fonts, but I'm not sure about which ones.

I still need to make Ruby's list and do some testing.

* Added a table for the multipliers on stats.

These are the values that are multiplied (then divided) when your Pokémon's at −6, −5, ..., +5, or +6 in battle.

* Filled in the rest of the allFonts List for EM.

07 = japan2
08 = japan1
02-05 = japan3

The game uses some font slots for multiple fonts. If you used \CC0608 before text, you would get glyphs from "graphics.text.font.short.characters" OR "graphics.text.font.japan1.characters", depending on what language you specified in your text string.

* Added the allFonts [[List]] for Ruby & Sapphire.

Many of the fonts in "data.fonts" are duplicates, albeit with minor numerical changes here and there.
I'd be fine with removing the '-01,' '-02,' etc. and just leaving duplicates of font names ('japan1,' 'japan1∼2,' etc.) if that's less confusing.

* Moved and standardized the stat changes table.

Changed the name and location, as well as adding another field to the table formatting.