More feature requests (#152)
* Updated the "jumpiftype" commands' type field.

* Added the other keypadIcons anchor.

The name was very similar in the decomps, but this anchor refers to a table instead of a sprite sheet.

* Added the HM logo image in tableReference.txt.

Thank you VersekrDarkYT for notifying us about this image in #feature-requests!

* Added Emerald's message box graphics anchor refs.

Thank you Lunos for the feature request! The word "noframe" for Emerald's message box identifies the message box graphics used for stuff like the "Would you like to save the game?" prompt. These type of message boxes are independent of the text frame configured by the player.

* Miscellaneous typo fixed.