finishing legendary event template
Legendary pokemon events each need 2 flags: 1 for "recent battle" and 1 for "capture". The legend template also includes an info textbox to show you how to reset the legendary battle if you want to. This is often done in a transition script, but you can do it anywhere. By default, HMA won't reset the battle for you: once you defeat the pokemon or run away, it'll stay gone until you reset it.