fix tileset repoint bug
when repointing tilesets, the tileset run gets a length of -1 if you try to create the run where there's only FF data. So instead of Cuting, we need to copy the data, uncompress it, resize it, and recompress it, and then repoint the original data to the new address... to get the data moved to the new point with a valid run. Once we change the data, we then need to updat ethe run since the length changed.