Removed the constant references for the XY coords.
We may need to do the same for the default BPEF and BPEI `.toml` files.

These addresses are actually part of thumb instructions and don't have anything to do with the spawn coordinates.

Actually, Emerald's code makes the player spawn at (−1, −1), which is a pair of invalid coordinates; therefore, the player actually spawns in the centre of the map. Another tricky thing is that there's no indication in the thumb code about those −1 coordinates — at least one we haven't found yet..