Partially overhauling Debug Asserts (#164)
* Changed the Debug flavor text. Added discord link.

Instead of "Show Full Message," I put "View Details" as previously discussed in a message chain in #feature−requests.

I also provided a Discord link in the same area as the stack trace so that people report debug asserts to #hma−bug−reports instead of dismissing them.

* Added the [Report] button & its functionality.

Basically, after clicking [Debug] on the debug assert message for the first time, you can click [Report] to open the Discord server (like you would on a crash dialogue). Do so after copying the stack trace.

More tweaks to this code may need to be done.

* Changed the way that HMA opens Discord in a DA.

I changed the method so that it's more online with previous lines of code in this file.