Don't starve the UI render thread during keyboard table scrolling
The new `WaitForRenderingAsync` method allows a ViewModel to check with the work dispatcher to make sure the render thread is caught up. We don't want to do this in all cases, but this gives us the ability to wait or skip events if rendering is taking too long, so that we don't starve the render thread.

For example, when using the `MoveSelectionStart` command, the ViewModel now checks if the rendering is caught up. If we're behind, the command request is simply ignored. If we're nearly caught up, it waits for rendering and then does the move work on a background task. If we're completely caught up, the work is still pushed to a background task, but is able to start running immediately.

If this becomes a performance issue and makes the app run too slowly, it'll be easy to just remove the await from the command.