Add python button capability for tables
Adding a field `name|python=function` to a table will add a button to the table. Clicking the button will run a python function from `hma.py`. The function is expected to have one element, which will be the current selected element of the table as a `ModelArrayElement`. This python code still has access to the entire rom and the print() function.
The button's text is based on the name of the field, and therefore cannot have a spaces or special characters, just letters / numbers.
The button will get tooltip that is based on the comment of the chosen function.