Added the other text strings.
"mainSpeech" is taken from the decomps, but it's not specific.

Also, I need to see if there is a table for these pointers as opposed to these text pointers existing in ASM. If the text strings are in a table, it makes implementing these text strings easier.

Note: The addresses in Emerald aren't actually the anchor refs b/c I can't generate them atm.