Added the "scripts.battle.badge.obey" category.
Credit to Axcellerator for finding these offsets.

For the life of me, I could not link the flags table to the other 4 new anchors.

Since all of these are part of a THUMB routine, nothing points to these individual anchors, so I had to put them here.

The 0badges to 6badges anchors link to each other, so that's why 3 of them do not have a 1 after their [level.] formatting.