use TransientModelDelta for removing/adding the dex info table
When adjusting the pokedex, all the same text is expected to be in the pokedex table, but in a different order. So the table gets cleared, the data gets changed, and the data gets re-added. This is fine, but it means that when the pointer is cleared, each text run gets removed since it (temporarily) has nothing pointing to it. This is a waste, since we expect to add pointers back by the end of the operation.

Introduce a new TransientModelDelta that is a type of NoDataChangeDeltaModel. This new type acts exactly like the NoDataChangeDeltaModel, except that it additionally signals to the model (if it cares) that the model doesn't need to worry about removing zero-pointer runs.