fix long-branch-link code
When pushing, the low registers are put at lower addresses, followed by lr. Pushing decreases sp.

When popping, the low registers taken from lower addresses, then pc is pulled from a higher address. Popping increases sp.

So we want to push r0/r1 and pop r0/pc, which means we want to manipulate r0, not r1. We only pop r1 so we have a space to pop pc from, not because we actually want to store the value.