add support for add/subtract macros in thumb code
See the new tests for examples of what you can do with this feature.
The first value in an = operation is allowed to be outside the range [0, 255]. All other values must be inside the range. You may add registers.
If you're doing += or -=, even the first value must be within the range [0, 255].
No support for parenthesis. High registers are supported only for addition or equality. So you could do:
`r8 = r9 + r10` or `r8 = r9`
but not
`r9 = 3` or `r9 = r2 + 3`