improve flyspot event UX
multiple users have gotten confused about why the flyspot event button is available sometimes, but not always. This change hopes to make that more clear.

Instead of the button only being there when it's useful, the button is always visible. However, the button is disabled (with a tooltip message) when it cannot be used.