button for making a flag for an OW
When working with hidesprite `var` instead of hidesprite `num`, the var may reference an OW that uses a hide/show flag that's never actually touched in a script. So you can't make the flag in a script first and then copy it into the map editor, it makes sense to be able to create a flag for the OW directly from within the map editor.