add capability of split-byte arguments for scripts
We want a macro that can "set the current move" for use in battle scripts, for setting how the animation should work. This needs to set a low-byte high-byte pair based on a move id, which requires 2 separate commands. In order for this macro to work, the code needs to be able to read the two split bytes as a single value, and write the single value as 2 split bytes. This is now supported.