Fixed the move effects [[List]] generation.
The existing code did not account for all of the new move effects, and the list added earlier was not completely accurate. Now, the move effect names _should_ be correct.

Some of the move effects (like Endure) are unused in HUBOL's _moves,_ but they're still used in some tables referenced by the vanilla AI.