Script discovery logic improvement
During the first pass, scripts that are detected that are earlier than the current script are given a length of 1 to prevent recursion.
The script parser can now produce a second pass on these addresses if they're recognized as scripts, to make sure that the full script gets concatenated rather than truncated to the first command.