add functionality for silencing notifications on child viewmodels
ViewModelCore can now use `SilencePropertyNotifications` to temporarily
silence notifications. These notifications are not deferred, they are silenced.

Furthermore, ViewModels can specify a list of child viewmodels that might need to be silenced. These children cannot be detected automatically, because some ViewModels hold references to peers as well as children.

When you call `SilencePropertyNotifications`, the ViewModel and all its registered children will be silenced until the end of the silence scope.