TableGroupPanel updates
After exploring other options of reducing property notifications or allowing async notifications, I've decided that the only real way to fix the table performance is to redo the UI layer. As such, I'm revisiting the TableGroupPanel.

I'll watch performance as I work to make sure it remains fast. But using a single control and modeling the interaction costs a lot of code, but hopefully not a lot of performance.