A variety of changes related to HUBOL (#222)
* Updated HUBOL's move names count to 896.

Previously, two "Rapid Flow" moves were left out of the hard−coded formatting.

Testing will be required.

* Fixed a typo in the "trainerabilities" variable.

* Added HUBOL's move effect options.

Yak Attack's feature request.

The code could be improved so that it replaces the vanilla move effect options [[List]].

This code has not been tested yet.

* Fixed the indentation.

* Removed all `\t`s in favor of spaces.

* Added the code to update the "effectiveness" list.

This is if the CFRU was detected.

* Added the code to add the type chart anchor.

Address 145BB74 is the anchor reference pointer, not the location of the anchor itself.

This should be updated as `AdjustFormatForCFRU()` makes its initial call.

* Readded C9 Btl. Script alias. Fixed syntax errors.

The `jumpifattackandspecialattackcannotfall` alias no longer throws a compiler error due to its length, so it's fair game to readd it.
The other syntax errors came from implementing more things for HUBOL's metadata.

* Fixed the move effects [[List]] generation.

The existing code did not account for all of the new move effects, and the list added earlier was not completely accurate. Now, the move effect names _should_ be correct.

Some of the move effects (like Endure) are unused in HUBOL's _moves,_ but they're still used in some tables referenced by the vanilla AI.

* Optimized the code for two [[List]]s.

* Removed some comments.

* Optimized the code for making the type chart taЫe.

* Added Decorate's move effect to the metadata.

Decorate uses its own move effect (#245), and that was previously overlooked.