Improve table tool loading performance by introducing delay
Don't have the table try to load unless it's stable for over half a second. This initial pause make the table slower, but prevents it from trying to load during scroll operations, which greatly increases the scrolling performance.

Also, don't make the table reload in response to self-edits. If the element isn't changing because the combobox is still open, leave it be until the filter closes.