Implemented a few more feature requests. (#266)
* Hold Effects Table: 1st param. references a table.

* Renamed "bag.inside3" to "bag.pokemart".

Requested by 1RWT16KU1D.

* Made a certain table's Trigger field a hex field.

I have not tested this yet. There's a chance this could lead to metadata irregularities.

* Added Emerald's vrsn. of the town map annotations.

Testing on a fresh ROM will commence shortly.

* Lowered the length of `scripts.text.destinations`.

The correct length is 8, not 14.

* Fixed a failed test involving an anchor change.

With `graphics.bag.inside.3.palette` and company renamed, the relevant assertion in `Upgrades.cs` should now succeed correctly. I just needed to supply the old names in one spot and the new names in another spot.

* Added the banlist for FireRed & LeafGreen.

This is for the battle tower.
Requested by Sylinuser.

* Added the same banlist for Ruby & Sapphire.