First partial draft of CrossTableDataViewPort
This ViewPort needs to be made editable and needs testing. But the general idea is that we can glom all the tables together and show them as rows in this ViewPort rather than using the table tool. Then we can farm out text, sprite, and palette data to the appropriate tools, so that we don't need to display everything in this one tab. Also need to figure something out for bitfields.